fix call url function
there was a recent commit (4073a22d5599bdf05589b91428ad6e16d7b6640d) that
changed what's returned in case of an http-error. Before, the response-body
was returned, after this change, only the return-code was returned.
(which is correct btw, according to rfc 2616: "The 304 response MUST NOT contain a
   message-body, and thus is always terminated by the first empty line
   after the header fields.")

Now we check if the response code is a 304 and set the response body to empty.